test: ユニットテストファイルを分割して整理
「クリア」ボタンのテストを、別ファイルに分割した。
