+++ src/app/components/BookmarkManager.test.tsx
@@ -774,33 +774,3 @@ describe("更新されたブックマークが、APIにPOSTで送られる。", 
     });
   });
 });
-
-describe("「クリア」ボタン", () => {
-  it("クリアボタンがクリックされたら、URLとタイトルテキストがクリアされる。", async () => {
-    fetchMock.mockResponseOnce(JSON.stringify(mockBookmarks));
-
-    await act(async () => {
-      render(
-        <MessageProvider>
-          <BmMessage />
-          <BookmarkManager />
-        </MessageProvider>
-      );
-    });
-
-    const urlInput = screen.getByRole("textbox", { name: "url" });
-    const titleInput = screen.getByRole("textbox", { name: "title" });
-    const clearButton = screen.getByRole("button", { name: "クリア" });
-
-    await act(async () => {
-      fireEvent.change(urlInput, { target: { value: "https://example.com" } });
-      fireEvent.change(titleInput, { target: { value: "Example Site" } });
-      fireEvent.click(clearButton);
-    });
-
-    await waitFor(() => {
-      expect(urlInput).toHaveValue("");
-      expect(titleInput).toHaveValue("");
-    });
-  });
-});
