+++ src/app/components/BookmarkManager/clear.test.tsx
@@ -1,0 +1,59 @@
+import "@testing-library/jest-dom";
+
+import fetchMock from "jest-fetch-mock";
+import React, { act } from "react";
+
+import { fireEvent, render, screen, waitFor } from "@testing-library/react";
+
+import { MessageProvider } from "../../contexts/MessageContext";
+import BmMessage from "../bmMessage";
+import { Bookmark, BookmarkManager } from "../BookmarkManager";
+
+const mockBookmarks: Bookmark[] = [
+  {
+    url: "https://github.com/kubotama/linkpage",
+    title: "kubotama/linkpage",
+  },
+  {
+    url: "https://www.google.com/",
+    title: "Google",
+  },
+  {
+    url: "https://mail.google.com",
+    title: "Gmail",
+  },
+  {
+    url: "https://www.amazon.co.jp/",
+    title: "Amazon",
+  },
+];
+
+describe("「クリア」ボタン", () => {
+  it("クリアボタンがクリックされたら、URLとタイトルテキストがクリアされる。", async () => {
+    fetchMock.mockResponseOnce(JSON.stringify(mockBookmarks));
+
+    await act(async () => {
+      render(
+        <MessageProvider>
+          <BmMessage />
+          <BookmarkManager />
+        </MessageProvider>
+      );
+    });
+
+    const urlInput = screen.getByRole("textbox", { name: "url" });
+    const titleInput = screen.getByRole("textbox", { name: "title" });
+    const clearButton = screen.getByRole("button", { name: "クリア" });
+
+    await act(async () => {
+      fireEvent.change(urlInput, { target: { value: "https://example.com" } });
+      fireEvent.change(titleInput, { target: { value: "Example Site" } });
+      fireEvent.click(clearButton);
+    });
+
+    await waitFor(() => {
+      expect(urlInput).toHaveValue("");
+      expect(titleInput).toHaveValue("");
+    });
+  });
+});
